requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.6
beautifulsoup4==4.9.3
boto3==1.17.57
botocore==1.20.57
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
defusedxml==0.7.1
distlib==0.3.1
Django==3.2.2
django-allauth==0.44.0
django-auto-prefetch==0.0.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-v5==1.0.2
django-contrib-comments==2.1.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-jsonview==2.0.0
django-markdownx==3.0.1
django-ranged-response==0.2.0
django-robots==4.0
django-simple-captcha==0.5.14
django-simple-history==3.0.0
django-storages==1.11.1
django-taggit==1.4.0
djangorestframework==3.12.4
docutils==0.17.1
drf-yasg==1.20.0
filelock==3.0.12
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
importlib-metadata==2.1.1
Impostor==2.0.5
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.3
Markdown==3.3.4
markdown2==2.4.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
promise==2.3
pycodestyle==2.7.0
pycparser==2.20
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.8.1
PyJWT==2.0.1
pymdown-extensions==8.1.1
pyparsing==2.4.7
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2021.1
readtime==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
Rx==1.6.1
s3transfer==0.4.2
singledispatch==3.6.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.4
zipp==3.4.1
